refactor(slack): separate monitor contracts from runtime policy
