+++ extensions/slack/src/monitor/context.ts
@@ -3,13 +3,11 @@ import type { App } from "@slack/bolt";
 import { formatAllowlistMatchMeta } from "openclaw/plugin-sdk/allow-from";
 import type { ChannelRuntimeSurface } from "openclaw/plugin-sdk/channel-contract";
 import type { PluginRuntime } from "openclaw/plugin-sdk/channel-core";
-import type { ChannelInboundTurnPlan } from "openclaw/plugin-sdk/channel-inbound";
 import type {
   OpenClawConfig,
   SlackReactionNotificationMode,
   SessionScope,
   DmPolicy,
-  GroupPolicy,
 } from "openclaw/plugin-sdk/config-contracts";
 import { createDedupeCache } from "openclaw/plugin-sdk/dedupe-runtime";
 import type { HistoryEntry } from "openclaw/plugin-sdk/reply-history";
@@ -26,13 +24,10 @@ import { renameSlackSession, setSlackSessionStatus } from "../session-status.js"
 import type { SlackMessageEvent } from "../types.js";
 import { createSlackAgentViewState } from "./agent-view-state.js";
 import { normalizeAllowList, normalizeAllowListLower, normalizeSlackSlug } from "./allow-list.js";
-import {
-  createSlackAssistantThreadContextStore,
-  type SlackAssistantThreadContext,
-} from "./assistant-thread-context.js";
-import type { SlackChannelConfigEntries } from "./channel-config.js";
+import { createSlackAssistantThreadContextStore } from "./assistant-thread-context.js";
 import { resolveSlackChannelConfig } from "./channel-config.js";
 import { normalizeSlackChannelType } from "./channel-type.js";
+import type { SlackMonitorContext, SlackChannelInfo } from "./context-types.js";
 import type { SlackIdentityHealth, SlackInstallationIdentity } from "./enterprise-install.js";
 import type { SlackEventScope } from "./event-scope.js";
 import { readLruMapEntry, writeLruMapEntry } from "./lru-map-cache.js";
@@ -42,31 +37,21 @@ import { createSlackRuntimeContextReader } from "./runtime-policy.js";
 import { isGovSlackClient } from "./slack-client-kind.js";
 import {
   type SlackSuggestedPromptsInput,
-  type SlackSuggestedPromptsOutcome,
   updateSlackSuggestedPrompts,
 } from "./suggested-prompts.js";
 import { createSlackSystemEventRouteResolver } from "./system-event-session.js";
 
+export type { SlackMonitorContext } from "./context-types.js";
 export { buildSlackAssistantThreadMetadata } from "./assistant-thread-context.js";
 export type { SlackAssistantThreadContext } from "./assistant-thread-context.js";
 export { normalizeSlackChannelType, resolveSlackChatType } from "./channel-type.js";
 export { DEFAULT_SLACK_SUGGESTED_PROMPTS } from "./suggested-prompts.js";
-
-type SlackChannelInfo = {
-  name?: string;
-  type?: SlackMessageEvent["channel_type"];
-  topic?: string;
-  purpose?: string;
-};
 
 type SlackChannelCacheEntry = {
   info: SlackChannelInfo;
   metadataLoaded: boolean;
 };
 
-type SlackUserInfo = { name?: string; imageUrl?: string; error?: unknown };
-type BuildChannelInboundContext =
-  typeof import("openclaw/plugin-sdk/channel-inbound").buildChannelInboundEventContext;
 const SLACK_CHANNEL_CACHE_MAX_ENTRIES = 1024;
 const SLACK_USER_CACHE_MAX_ENTRIES = 2048;
 const SLACK_AVATAR_CACHE_MAX_ENTRIES = 128;
@@ -77,115 +62,6 @@ const SLACK_AVATAR_SSRF_POLICY = {
 };
 const SLACK_CHANNEL_DENIAL_WARNING_TTL_MS = 5 * 60_000;
 const SLACK_CHANNEL_DENIAL_WARNING_MAX_ENTRIES = 1024;
-
-export type SlackMonitorContext = {
-  cfg: OpenClawConfig;
-  readRuntimeContext: () => Promise<SlackMonitorContext>;
-  isRuntimePolicyCurrent: () => boolean;
-  accountId: string;
-  botToken: string;
-  app: App;
-  runtime: RuntimeEnv;
-  channelRuntime?: ChannelRuntimeSurface;
-  buildContext?: BuildChannelInboundContext;
-  dispatchReplyFromConfig?: ChannelInboundTurnPlan["dispatchReplyFromConfig"];
-
-  botUserId: string;
-  botId?: string;
-  identityHealth: SlackIdentityHealth;
-  teamId: string;
-  apiAppId: string;
-  installationIdentity: SlackInstallationIdentity;
-
-  historyLimit: number;
-  dmHistoryLimit: number;
-  channelHistories: Map<string, HistoryEntry[]>;
-  sessionScope: SessionScope;
-  mainKey: string;
-
-  dmEnabled: boolean;
-  dmPolicy: DmPolicy;
-  allowFrom: string[];
-  allowNameMatching: boolean;
-  groupDmEnabled: boolean;
-  groupDmChannels: string[];
-  defaultRequireMention: boolean;
-  channelsConfig?: SlackChannelConfigEntries;
-  channelsConfigKeys: string[];
-  groupPolicy: GroupPolicy;
-  useAccessGroups: boolean;
-  reactionMode: SlackReactionNotificationMode;
-  reactionAllowlist: Array<string | number>;
-  replyToMode: "off" | "first" | "all" | "batched";
-  threadHistoryScope: "thread" | "channel";
-  threadInheritParent: boolean;
-  slashCommand: Required<import("openclaw/plugin-sdk/config-contracts").SlackSlashCommandConfig>;
-  textLimit: number;
-  typingReaction: string;
-  mediaMaxBytes: number;
-
-  logger: ReturnType<typeof getChildLogger>;
-  shouldDropMismatchedSlackEvent: (body: unknown) => boolean;
-  resolveSlackSystemEventRoute: (params: {
-    channelId?: string | null;
-    channelType?: string | null;
-    senderId?: string | null;
-    threadTs?: string | null;
-    eventScope?: SlackEventScope;
-  }) => { agentId: string; sessionKey: string };
-  isChannelAllowed: (params: {
-    teamId?: string;
-    channelId?: string;
-    channelName?: string;
-    channelType?: SlackMessageEvent["channel_type"];
-  }) => boolean;
-  resolveChannelName: (
-    channelId: string,
-    eventScope?: SlackEventScope,
-  ) => Promise<SlackChannelInfo>;
-  /** Records authoritative event-carried channel type in the channel metadata cache. */
-  rememberSlackChannelType: (
-    channelId: string | null | undefined,
-    channelType: string | null | undefined,
-    eventScope?: SlackEventScope,
-  ) => void;
-  /** Reads event-carried channel type when Slack omits it from later bot/edit/delete events. */
-  recallSlackChannelType: (
-    channelId: string | null | undefined,
-    eventScope?: SlackEventScope,
-  ) => SlackMessageEvent["channel_type"] | undefined;
-  resolveUserName: (userId: string, eventScope?: SlackEventScope) => Promise<SlackUserInfo>;
-  resolveUserAvatar: (userId: string, eventScope?: SlackEventScope) => string | undefined;
-  setSlackSessionStatus: (params: {
-    channelId: string;
-    threadTs?: string;
-    status: "processing" | "active" | "suspended";
-    title?: string;
-    eventScope?: SlackEventScope;
-  }) => Promise<void>;
-  recordSlackSessionTitle: (params: {
-    channelId: string;
-    threadTs: string;
-    title: string;
-    eventScope?: SlackEventScope;
-  }) => void;
-  getSlackAssistantThreadContext: (
-    channelId: string | undefined,
-    threadTs: string | undefined,
-    eventScope?: SlackEventScope,
-  ) => SlackAssistantThreadContext | undefined;
-  saveSlackAssistantThreadContext: (
-    context: Omit<SlackAssistantThreadContext, "updatedAt">,
-    eventScope?: SlackEventScope,
-  ) => void;
-  setSlackSuggestedPrompts: (
-    params: SlackSuggestedPromptsInput,
-  ) => Promise<SlackSuggestedPromptsOutcome>;
-  recordSlackAgentView: () => Promise<void>;
-  isSlackAgentView: () => Promise<boolean>;
-  recordSlackManagedViewThread: (channelId: string, threadTs: string) => Promise<void>;
-  isSlackManagedViewThread: (channelId: string, threadTs: string) => Promise<boolean>;
-};
 
 export function createSlackMonitorContext(params: {
   cfg: OpenClawConfig;
