+++ extensions/slack/src/monitor/events/interactions.shortcuts.ts
@@ -29,27 +29,27 @@ async function handleSlackShortcut(params: {
 }): Promise<void> {
   const { ack, body } = params.args;
   await ack();
-  params = { ...params, ctx: await params.ctx.readRuntimeContext() };
+  const runtimeContext = await params.ctx.readRuntimeContext();
   const eventScope = resolveSlackListenerEventScope({
-    identity: params.ctx.installationIdentity,
+    identity: runtimeContext.installationIdentity,
     body,
     context: params.args.context,
     client: params.args.client,
-    clientOptions: params.ctx.app.webClientOptions,
-    onDrop: (reason) => params.ctx.runtime.log?.(`slack:interaction drop shortcut ${reason}`),
+    clientOptions: runtimeContext.app.webClientOptions,
+    onDrop: (reason) => runtimeContext.runtime.log?.(`slack:interaction drop shortcut ${reason}`),
   });
   if (eventScope === null) {
     return;
   }
-  if (params.ctx.shouldDropMismatchedSlackEvent?.(body)) {
-    params.ctx.runtime.log?.("slack:interaction drop shortcut payload (mismatched app/team)");
+  if (runtimeContext.shouldDropMismatchedSlackEvent?.(body)) {
+    runtimeContext.runtime.log?.("slack:interaction drop shortcut payload (mismatched app/team)");
     return;
   }
 
   const callbackId = body.callback_id?.trim();
   const userId = body.user?.id?.trim();
   if (!callbackId || !userId) {
-    params.ctx.runtime.log?.("slack:interaction drop shortcut reason=invalid-payload");
+    runtimeContext.runtime.log?.("slack:interaction drop shortcut reason=invalid-payload");
     return;
   }
   params.trackEvent?.();
@@ -58,14 +58,14 @@ async function handleSlackShortcut(params: {
   const messageBody = isMessageShortcut ? body : undefined;
   const channelId = messageBody?.channel.id?.trim() || undefined;
   if (isMessageShortcut && !channelId) {
-    params.ctx.runtime.log?.(
+    runtimeContext.runtime.log?.(
       `slack:interaction drop shortcut callback=${callbackId} user=${userId} reason=missing-channel`,
     );
     return;
   }
   const threadTs = messageBody ? resolveMessageThreadTs(messageBody) : undefined;
   const auth = await authorizeSlackSystemEventSender({
-    ctx: params.ctx,
+    ctx: runtimeContext,
     eventScope,
     senderId: userId,
     channelId,
@@ -74,7 +74,7 @@ async function handleSlackShortcut(params: {
     interactiveEvent: true,
   });
   if (!auth.allowed) {
-    params.ctx.runtime.log?.(
+    runtimeContext.runtime.log?.(
       `slack:interaction drop shortcut callback=${callbackId} user=${userId} reason=${auth.reason ?? "unauthorized"}`,
     );
     return;
@@ -104,7 +104,7 @@ async function handleSlackShortcut(params: {
     messageText: messageBody?.message.text,
     responseUrl: messageBody?.response_url,
   };
-  const route = params.ctx.resolveSlackSystemEventRoute({
+  const route = runtimeContext.resolveSlackSystemEventRoute({
     channelId,
     channelType: auth.channelType,
     senderId: userId,
@@ -123,7 +123,7 @@ async function handleSlackShortcut(params: {
     .filter(Boolean)
     .join(":");
 
-  params.ctx.runtime.log?.(
+  runtimeContext.runtime.log?.(
     `slack:interaction ${interactionType} callback=${callbackId} user=${userId} channel=${channelId ?? "direct"}`,
   );
   const queued = enqueueRoutedSystemEvent(params.formatSystemEvent(eventPayload), route, {
@@ -131,7 +131,7 @@ async function handleSlackShortcut(params: {
     deliveryContext: {
       channel: "slack",
       to: deferredTarget.target,
-      accountId: params.ctx.accountId,
+      accountId: runtimeContext.accountId,
       threadId: threadTs,
     },
   });
