+++ extensions/slack/src/monitor/events/reactions.ts
@@ -56,21 +56,24 @@ export function registerSlackReactionEvents(params: {
     eventId: string,
   ) => {
     try {
-      const ctx = await params.ctx.readRuntimeContext();
+      const runtimeContext = await params.ctx.readRuntimeContext();
       const item = event.item;
       if (!item || item.type !== "message") {
         return;
       }
-      if (ctx.reactionMode === "off") {
+      if (runtimeContext.reactionMode === "off") {
         return;
       }
-      if (ctx.reactionMode === "own" && (!ctx.botUserId || event.item_user !== ctx.botUserId)) {
+      if (
+        runtimeContext.reactionMode === "own" &&
+        (!runtimeContext.botUserId || event.item_user !== runtimeContext.botUserId)
+      ) {
         return;
       }
       trackEvent?.();
 
       const ingressContext = await authorizeAndResolveSlackSystemEventContext({
-        ctx,
+        ctx: runtimeContext,
         senderId: event.user,
         channelId: item.channel,
         eventKind: "reaction",
@@ -89,7 +92,7 @@ export function registerSlackReactionEvents(params: {
       const [actorInfo, authorInfo] = await Promise.all([actorInfoPromise, authorInfoPromise]);
       if (
         !shouldEmitSlackReactionNotification({
-          ctx,
+          ctx: runtimeContext,
           event,
           eventScope,
           actorName: actorInfo?.name,
